+++ mutex4.c
@@ -49,7 +49,7 @@ static pthread_mutex_t mutex1;
  
 void * unlocker(void * arg)
 {
-  int expectedResult = (int) arg;
+  int expectedResult = (long) arg;
 
   wasHere++;
   assert(pthread_mutex_unlock(&mutex1) == expectedResult);
